Change the location of psrcat to /opt/psrcat
There was a subtle bug that prevented psrcat from working when called from the /software directory (due to singularity binding).
This has now been changed to /opt/psrcat inside the container.
I also renamed the parameters to remove the psrcat prefix
since as Dane pointed out ... it is a bit redundant.